+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 Version 1.7.3 (26-05-28)
 ------------------------
 * added caching for audmodel
+* add layer selection for audmodel
 
 Version 1.7.2 (26-05-28)
 ------------------------